fix: #채팅,회원 장원태 [240206] - MemberDto에서 usercode를 int로 수정 - 채팅의 별점 모달 구성중

+++ src/main/java/mypage/repository/MemberDaoInter.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import mypage.data.MemberDto;
 
-public interface MemberDaoInter extends JpaRepository<MemberDto, Long>{
+public interface MemberDaoInter extends JpaRepository<MemberDto, Integer>{
 }
